♻️ refactor: 별점 렌더링 함수 util화
